+++ tests/bootstrap.php
@@ -10,12 +10,7 @@
  * @link    http://nikolazeravcic.iz.rs Personal site
  */
 
+// Needed for isolated tests
+//require __DIR__ . '/../vendor/autoload.php';
 
-error_reporting(E_ALL);
-
-?>
-
-<h1>TTDs By John Cleary</h1>
-
-<p><em>You can set testing code here</em></p>
-
+require 'vendor/autoload.php';