+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All Notable changes to `bytenet\laravel-admn-base` project will be documented in
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased] - 2016-12-27
+### Changed
+- tests/bootstrap.php
+- .travis.yml file
 
 ## 0.1.0 - 2016-12-27
 ### Added
 -  Day-01 - SRC and TEST
+
+
+[Unreleased]: https://github.com/zeravcic/tdd_by_john-cleary/compare/v0.1.0...HEAD